Move amundsen queries
