+++ ingestion/src/metadata/ingestion/source/metadata/amundsen/metadata.py
@@ -66,15 +66,15 @@ from metadata.ingestion.ometa.client_utils import get_chart_entities_from_id
 from metadata.ingestion.ometa.ometa_api import OpenMetadata
 from metadata.ingestion.source.connections import get_connection
 from metadata.ingestion.source.database.column_type_parser import ColumnTypeParser
-from metadata.utils import fqn
-from metadata.utils.helpers import get_standard_chart_type
-from metadata.utils.logger import ingestion_logger
-from metadata.utils.metadata_service_helper import SERVICE_TYPE_MAPPER
-from metadata.utils.sql_queries import (
+from metadata.ingestion.source.metadata.amundsen.queries import (
     NEO4J_AMUNDSEN_DASHBOARD_QUERY,
     NEO4J_AMUNDSEN_TABLE_QUERY,
     NEO4J_AMUNDSEN_USER_QUERY,
 )
+from metadata.utils import fqn
+from metadata.utils.helpers import get_standard_chart_type
+from metadata.utils.logger import ingestion_logger
+from metadata.utils.metadata_service_helper import SERVICE_TYPE_MAPPER
 
 logger = ingestion_logger()
