+++ ingestion/src/metadata/ingestion/source/metadata/amundsen/queries.py
@@ -1,0 +1,115 @@
+#  Copyright 2021 Collate
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#  http://www.apache.org/licenses/LICENSE-2.0
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+"""
+SQL Queries used during ingestion
+"""
+
+import textwrap
+
+NEO4J_AMUNDSEN_TABLE_QUERY = textwrap.dedent(
+    """
+        MATCH (db:Database)<-[:CLUSTER_OF]-(cluster:Cluster)
+        <-[:SCHEMA_OF]-(schema:Schema)<-[:TABLE_OF]-(table:Table)
+        OPTIONAL MATCH (table)-[:DESCRIPTION]->(table_description:Description)
+        OPTIONAL MATCH (schema)-[:DESCRIPTION]->(schema_description:Description)
+        OPTIONAL MATCH (table)-[:DESCRIPTION]->(prog_descs:Programmatic_Description)
+        WITH db, cluster, schema, schema_description, table, table_description,
+        COLLECT(prog_descs.description) as programmatic_descriptions
+        OPTIONAL MATCH (table)-[:TAGGED_BY]->(tags:Tag) WHERE tags.tag_type='default'
+        WITH db, cluster, schema, schema_description, table, table_description, programmatic_descriptions,
+        COLLECT(DISTINCT tags.key) as tags
+        OPTIONAL MATCH (table)-[:HAS_BADGE]->(badges:Badge)
+        WITH db, cluster, schema, schema_description, table, table_description, programmatic_descriptions, tags,
+        COLLECT(DISTINCT badges.key) as badges
+        OPTIONAL MATCH (table)-[read:READ_BY]->(user:User)
+        WITH db, cluster, schema, schema_description, table, table_description, programmatic_descriptions, tags, badges,
+        SUM(read.read_count) AS total_usage,
+        COUNT(DISTINCT user.email) as unique_usage
+        OPTIONAL MATCH (table)-[:COLUMN]->(col:Column)
+        OPTIONAL MATCH (col)-[:DESCRIPTION]->(col_description:Description)
+        WITH db, cluster, schema, schema_description, table, table_description, tags, badges, total_usage, unique_usage,
+        programmatic_descriptions,
+        COLLECT(col.name) AS column_names, COLLECT(col_description.description) AS column_descriptions,
+        COLLECT(col.col_type) AS column_types
+        OPTIONAL MATCH (table)-[:LAST_UPDATED_AT]->(time_stamp:Timestamp)
+        RETURN db.name as database, cluster.name AS cluster, schema.name AS schema,
+        schema_description.description AS schema_description,
+        table.name AS name, table.key AS key, table_description.description AS description,
+        time_stamp.last_updated_timestamp AS last_updated_timestamp,
+        column_names,
+        column_descriptions,
+        column_types,
+        total_usage,
+        unique_usage,
+        tags,
+        badges,
+        programmatic_descriptions
+        ORDER BY table.name;
+        """
+)
+
+NEO4J_AMUNDSEN_USER_QUERY = textwrap.dedent(
+    """
+        MATCH (user:User)
+        OPTIONAL MATCH (user)-[read:READ]->(a)
+        OPTIONAL MATCH (user)-[own:OWNER_OF]->(b)
+        OPTIONAL MATCH (user)-[follow:FOLLOWED_BY]->(c)
+        OPTIONAL MATCH (user)-[manage_by:MANAGE_BY]->(manager)
+        with user, a, b, c, read, own, follow, manager
+        where user.full_name is not null
+        return user.email as email, user.first_name as first_name, user.last_name as last_name,
+        user.full_name as full_name, user.github_username as github_username, user.team_name as team_name,
+        user.employee_type as employee_type, manager.email as manager_email,
+        user.slack_id as slack_id, user.is_active as is_active, user.role_name as role_name,
+        REDUCE(sum_r = 0, r in COLLECT(DISTINCT read)| sum_r + r.read_count) AS total_read,
+        COLLECT(DISTINCT b) as entities_owned,
+        count(distinct b) as total_own,
+        count(distinct c) AS total_follow
+        order by user.email
+        """
+)
+
+NEO4J_AMUNDSEN_DASHBOARD_QUERY = textwrap.dedent(
+    """
+         MATCH (dashboard:Dashboard)
+         MATCH (dashboard)-[:DASHBOARD_OF]->(dbg:Dashboardgroup)
+         MATCH (dbg)-[:DASHBOARD_GROUP_OF]->(cluster:Cluster)
+         OPTIONAL MATCH (dashboard)-[:DESCRIPTION]->(db_descr:Description)
+         OPTIONAL MATCH (dbg)-[:DESCRIPTION]->(dbg_descr:Description)
+         OPTIONAL MATCH (dashboard)-[:EXECUTED]->(last_exec:Execution)
+         WHERE split(last_exec.key, '/')[5] = '_last_successful_execution'
+         OPTIONAL MATCH (dashboard)-[read:READ_BY]->(user:User)
+         WITH dashboard, dbg, db_descr, dbg_descr, cluster, last_exec, SUM(read.read_count) AS total_usage
+         OPTIONAL MATCH (dashboard)-[:HAS_QUERY]->(query:Query)-[:HAS_CHART]->(chart:Chart)
+         WITH dashboard, dbg, db_descr, dbg_descr, cluster, last_exec, COLLECT(DISTINCT query.name) as query_names,
+         COLLECT(chart.name) as chart_names,
+         COLLECT(chart.id) as chart_ids,
+         COLLECT(chart.url) as chart_urls,
+         COLLECT(chart.type) as chart_types,
+         total_usage
+         OPTIONAL MATCH (dashboard)-[:TAGGED_BY]->(tags:Tag) WHERE tags.tag_type='default'
+         WITH dashboard, dbg, db_descr, dbg_descr, cluster, last_exec, query_names, chart_names, chart_ids, chart_urls,
+         chart_types, total_usage,
+         COLLECT(DISTINCT tags.key) as tags
+         OPTIONAL MATCH (dashboard)-[:HAS_BADGE]->(badges:Badge)
+         WITH  dashboard, dbg, db_descr, dbg_descr, cluster, last_exec, query_names, chart_names, chart_ids, chart_urls,
+         chart_types, total_usage, tags,
+         COLLECT(DISTINCT badges.key) as badges
+         RETURN dbg.name as group_name, dashboard.name as name, cluster.name as cluster,
+         coalesce(db_descr.description, '') as description,
+         coalesce(dbg.description, '') as group_description, dbg.dashboard_group_url as group_url,
+         dashboard.dashboard_id as dashboard_id,
+         dashboard.dashboard_url as url, dashboard.key as uri,
+         split(dashboard.key, '_')[0] as product, toInteger(last_exec.timestamp) as last_successful_run_timestamp,
+         query_names, chart_names, chart_ids, chart_urls, chart_types, total_usage, tags, badges
+         order by dbg.name
+        """
+)
